# src/swagger.yaml
openapi: 3.0.0
info:
  title: GMR_SOCKET.IO_API DOCS
  version: 1.0.0
servers:
  - url: http://localhost:3001
...
security:
  - BearerAuth: []
  - ClientIdHeader: []
paths:
...
  /user:
    get:
      tags:
        - User
      summary: Get all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '400':
          description: Failed to retrieve users
...
  /user/{id}:
    put:
      tags:
        - User
      summary: Update a user category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Category updated successfully
        '400':
          description: Update failed
components:
  schemas:
    UserDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    ClientIdHeader:
      type: apiKey
      in: header
      name: clientid